Restructured so that only the required functions are exposed to the end user. AMAAS-385
